[+]update package [+]convert js files to jsx [+]add webpack and commands for working with webpack [+]update babelrc [-]delete build from rep
