+++ .babelrc
@@ -1,13 +1,13 @@
 {
   "presets": ["@babel/preset-env","@babel/preset-react"],
   "plugins": [
-    ["@babel/transform-runtime",
+    [
+      "@babel/plugin-transform-runtime",
       {
         "regenerator": true
       }
     ],
     ["@babel/plugin-proposal-decorators", { "legacy": true }],
-    "transform-class-properties",
-    "transform-object-rest-spread"
+    ["@babel/plugin-proposal-class-properties", { "loose": true }]
   ]
 }
