+++ .gitignore
@@ -1,7 +1,3 @@
-node_modules/
-.cache/
-*.map
 .idea
-.eslintrc
-package-lock.json
-src/tests/
+node_modules
+src/build
