+++ package.json
@@ -1,47 +1,70 @@
 {
-  "name": "react-extension",
-  "version": "2.1.0",
+  "name": "jdi-react-extension",
+  "version": "1.0.0",
   "description": "jdi react extension",
-  "main": "src/js/main.js",
   "scripts": {
-    "build": "parcel build src/js/main.js -d src/build/ -o main.js",
-    "watch": "parcel watch src/js/main.js -d src/build/ -o main.js",
-    "test": "jest"
+    "start": "node server.js",
+    "webpack": "webpack",
+    "build": "webpack"
   },
+  "main": "src/js/main.js",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/mobxjs/mobx-react-boilerplate.git"
+  },
+  "keywords": [
+    "react",
+    "reactjs",
+    "boilerplate",
+    "mobx",
+    "starter-kit"
+  ],
   "author": "anisa",
   "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/mobxjs/mobx/issues"
+  },
+  "homepage": "http://mobxjs.github.com/mobx",
   "dependencies": {
     "classnames": "^2.2.6",
-    "mobx": "^5.6.0",
-    "mobx-react": "^5.3.6",
-    "parcel-bundler": "^1.6.2",
-    "primer": "^10.9.0",
-    "react": "^16.6.1",
-    "react-dom": "^16.2.0",
+    "mobx": "^5.15.4",
+    "mobx-react": "^6.2.2",
+    "mobx-react-devtools": "^6.1.1",
+    "parcel-bundler": "^1.12.4",
+    "primer": "^11.0.0",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "react-file-reader": "^1.1.4",
-    "react-jss": "^8.6.1",
+    "react-jss": "^10.1.1",
     "react-loading": "^2.0.3",
-    "react-responsive": "^6.0.1",
-    "react-select": "^2.2.0",
-    "simple-crypto-js": "^2.0.2"
+    "react-responsive": "^8.0.3",
+    "react-select": "^3.1.0",
+    "simple-crypto-js": "^2.2.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.3.4",
-    "@babel/plugin-proposal-class-properties": "^7.1.0",
-    "@babel/plugin-proposal-decorators": "^7.3.0",
-    "@babel/plugin-transform-runtime": "^7.3.4",
-    "@babel/preset-env": "^7.3.4",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/runtime": "^7.3.4",
-    "babel-jest": "^23.6.0",
+    "@babel/core": "^7.9.0",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/plugin-proposal-decorators": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.9.0",
+    "@babel/preset-env": "^7.9.5",
+    "@babel/preset-react": "^7.9.4",
+    "@babel/runtime": "^7.9.2",
+    "babel-jest": "^25.4.0",
+    "babel-loader": "^8.1.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.5",
     "babel-plugin-transform-object-rest-spread": "^7.0.0-beta.3",
-    "file-saver": "^2.0.0",
-    "jest": "^23.6.0",
-    "react-test-renderer": "^16.7.0"
-  },
-  "resolutions": {
-    "deasync": "0.1.16"
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-mobx": "^2.0.0",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-1": "^6.24.1",
+    "file-saver": "^2.0.2",
+    "jest": "^25.4.0",
+    "json-loader": "^0.5.7",
+    "react-test-renderer": "^16.13.1",
+    "webpack": "^4.42.1",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3"
   }
 }
