+++ src/js/blocks/generate/left/GenerateBlock.jsx
