+++ src/js/blocks/generate/left/GenerateSettings.jsx
