+++ src/js/blocks/generate/right/GeneralSettings.jsx
