+++ src/js/blocks/generate/right/GenerateResults.jsx
