+++ src/js/blocks/log/LogComponent.jsx
