+++ src/js/blocks/rules/left/RulesBlock.jsx
