+++ src/js/components/Infinity/Infinity.jsx
