+++ src/js/components/Label/Label.jsx
