+++ src/js/json/JavaJDILightTemplate.jsx
@@ -17,14 +17,10 @@ public class {{siteName}} {
     public static {{type}} {{name}};`,
 	"page" : `package {{package}}.pages;
 
-import com.epam.jdi.light.elements.pageobjects.annotations.locators.*;
+import com.epam.jdi.light.elements.pageobjects.annotations.simple.UI;
 import com.epam.jdi.light.elements.composite.*;
-import com.epam.jdi.light.elements.complex.*;
-import com.epam.jdi.light.elements.common.*;
-import com.epam.jdi.light.elements.complex.dropdown.*;
-import com.epam.jdi.light.elements.complex.table.*;
-import com.epam.jdi.light.ui.html.elements.complex.*;
-import com.epam.jdi.light.ui.html.elements.common.*;
+import com.epam.jdi.light.ui.html.complex.*;
+import com.epam.jdi.light.ui.html.common.*;
 import {{package}}.sections.*;
 
 public class {{type}} extends WebPage {
@@ -32,20 +28,16 @@ public class {{type}} extends WebPage {
 }`,
 	"pageElementCss" : `    @UI("{{locator}}") public {{type}} {{name}};`,
 	"pageElementXPath" : `    @UI("{{locator}}") public {{type}} {{name}};`,
-	"pageElementComplex" : `    @JDropdown({{locators}}) 
+	"pageElementComplex" : `    @J{{type}}({{locators}}) 
 	public {{type}} {{name}};`,
 	"locatorCss" : `{{type}} = "{{locator}}",`,
 	"locatorXPath" : `{{type}} = "{{locator}}",`,
 	"section" : `package {{package}}.sections;
 
-import com.epam.jdi.light.elements.pageobjects.annotations.locators.*;
+import com.epam.jdi.light.elements.pageobjects.annotations.simple.UI;
 import com.epam.jdi.light.elements.composite.*;
-import com.epam.jdi.light.elements.complex.*;
-import com.epam.jdi.light.elements.common.*;
-import com.epam.jdi.light.elements.complex.dropdown.*;
-import com.epam.jdi.light.elements.complex.table.*;
-import com.epam.jdi.light.ui.html.elements.complex.*;
-import com.epam.jdi.light.ui.html.elements.common.*;
+import com.epam.jdi.light.ui.html.complex.*;
+import com.epam.jdi.light.ui.html.common.*;
 
 public class {{type}} extends Section {
 {{elements}}	
@@ -53,14 +45,10 @@ public class {{type}} extends Section {
 
 	"form" : `package {{package}}.sections;
 
-import com.epam.jdi.light.elements.pageobjects.annotations.locators.*;
+import com.epam.jdi.light.elements.pageobjects.annotations.simple.UI;
 import com.epam.jdi.light.elements.composite.*;
-import com.epam.jdi.light.elements.complex.*;
-import com.epam.jdi.light.elements.common.*;
-import com.epam.jdi.light.elements.complex.dropdown.*;
-import com.epam.jdi.light.elements.complex.table.*;
-import com.epam.jdi.light.ui.html.elements.complex.*;
-import com.epam.jdi.light.ui.html.elements.common.*;
+import com.epam.jdi.light.ui.html.complex.*;
+import com.epam.jdi.light.ui.html.common.*;
 import {{package}}.entities.*;
 
 public class {{type}} extends Form<{{data}}> {
