+++ src/js/json/settings.jsx
@@ -1,13 +1,13 @@
 const Languages = [
 	{ value: '.java', label: 'Java' },
-	{ value: '.cs', label: 'C#' }
+	{ value: '.cs', label: 'C#' },
 ];
 
 const Frameworks = [
 	{ value: 'jdiLight', label: 'JDI Light' },
 	{ value: 'jdiUI', label: 'JDI UI' },
 	{ value: 'selenium', label: 'Selenium' },
-	{ value: 'selenide', label: 'Selenide' }
+	{ value: 'selenide', label: 'Selenide' },
 ];
