--- src/js/json/siteUrls.js
@@ -1,9 +1,0 @@
-const SiteUrls = {
-	urlList: [
-		'https://jdi-testing.github.io/jdi-light/index.html',
-		'https://jdi-testing.github.io/jdi-light/contacts.html',
-		'https://jdi-testing.github.io/jdi-light/metals-colors.html'
-	]
-};
-
-export { SiteUrls };
