+++ src/js/json/siteUrls.jsx
@@ -1,0 +1,8 @@
+const SiteUrls = {
+	urlList: [
+		'https://epam.github.io/JDI/index.html',
+		'https://epam.github.io/JDI/contacts.html',
+	]
+};
+
+export { SiteUrls };
