+++ src/js/models/Log.jsx
