+++ src/js/panel.js
@@ -2,6 +2,6 @@ window.onload = () => {
 	chrome.devtools.panels.create("JDN",
 		"icon.png",
 		"panel.html",
-		function(panel) {}
+		(panel) => {console.log(panel)}
 	);
 };
